Time Changes to i64

+++ wServer/networking/packets/incoming/AoeAck.cs
@@ -4,7 +4,7 @@ namespace wServer.networking.packets.incoming;
 
 public class AoeAck : IncomingMessage
 {
-    public int Time { get; set; }
+    public long Time { get; set; }
     public Position Position { get; set; }
 
     public override C2SPacketId C2SId => C2SPacketId.AoeAck;
@@ -12,7 +12,7 @@ public class AoeAck : IncomingMessage
 
     protected override void Read(NReader rdr)
     {
-        Time = rdr.ReadInt32();
+        Time = rdr.ReadInt64();
         Position = Position.Read(rdr);
     }
     protected override void Write(NWriter wtr)
